Extract dedent(_:) into CDMarkdownParser+Dedent.swift
CDMarkdownParser.swift exceeded SwiftLint's file_length (750) and
type_body_length (500) limits once the dedent helper was added. Splitting
it into its own extension file matches this codebase's established
Type+Concern.swift convention for extension files.
